+++ src/NeuralBotInstance.h
@@ -40,6 +40,7 @@ public:
     void OnPlayerCreatureKill(Creature* killed = nullptr);
     void OnPlayerLearnSpell(uint32 spellId);
     void ReviveIfDead();
+    void StageEpisodeStart();
     void ProcessBotPackets();
     void OnWorldUpdate(uint32 diff);